devtools/client/shared/components/reps/test/mochitest/test_reps_nan.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test NaN rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - NaN</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
"use strict";

window.onload = Task.async(function* () {
  const {
    REPS,
    getRep,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { NaNRep } = REPS;

  try {
    yield testNaN();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testNaN() {
    const stub = {
      type: "NaN"
    };
    is(getRep(stub), NaNRep.rep, "Rep correctly selects NaN Rep");

    const renderedComponent = renderComponent(NaNRep.rep, {object: stub});
    is(renderedComponent.textContent, "NaN", "NaN rep has expected text content");
  }
});
</script>
</pre>
</body>
</html>
